+++ src/alignment/mod.rs
@@ -23,6 +23,7 @@ mod errors;
 mod pairwise;
 mod profile;
 mod profile_set;
+mod sneaky_snake;
 mod std_traits;
 mod types;
 
@@ -30,6 +31,7 @@ pub use errors::*;
 pub use pairwise::*;
 pub use profile::*;
 pub use profile_set::*;
+pub use sneaky_snake::*;
 pub use types::*;
 
 /// A trait implemented on structs containing data fields related to a query